requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
Flask-Caching==2.0.2
Werkzeug==2.3.7
gunicorn==21.2.0